:memo: examples
