--- aksel.nav.no/website/pages/eksempler/info-card/title-only.tsx
@@ -1,25 +1,0 @@
-import { InfoCard } from "@navikt/ds-react";
-import { withDsExample } from "@/web/examples/withDsExample";
-
-const Example = () => {
-  return (
-    <InfoCard data-color="info">
-      <InfoCard.Header>
-        <InfoCard.Title>Fremhevet statisk informasjon</InfoCard.Title>
-      </InfoCard.Header>
-    </InfoCard>
-  );
-};
-
-// EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
-export default withDsExample(Example);
-
-/* Storybook story */
-export const Demo = {
-  render: Example,
-};
-
-export const args = {
-  index: 6,
-  desc: "For korte meldinger kan man velge å kun vise tittel i InfoCard. Title er en `h2`-tagg som standard, husk å justere semantikken etter behov.",
-};
